feat: changed login method to update user status

See merge request cs309/2026spring/4_sh_2!16

+++ Backend/4_sh_2_backend_project/src/main/java/onetoone/Users/UserController.java
@@ -88,8 +88,14 @@ public class UserController {
     @DeleteMapping(path = "/users/{id}")
         // delete the user that matches the id
     String deleteUser(@PathVariable int id) {
+        if (userRepository.findById(id) == null) {
+            return failure;
+        }
         userRepository.deleteById(id);
-        return success;
+        if (userRepository.findById(id) == null) {
+            return success;
+        }
+        return failure;
     }
 
 
@@ -107,6 +113,8 @@ public class UserController {
         }
 
         // otherwise, login successful, and return the user object to the frontend
+        user.setUserStatus(UserStatus.ONLINE);
+        userRepository.save(user);
         return user;
     }
